fixes & enhancements
Done_DestroyByBoundary.cs now advances level when it destroys the boss
that exits the bottom of the screen, to catch the case where player
doesn't kill the boss in time.

Added health points to player, enemies, and asteroids.
Damage is taken upon collision with bullets and other fire.

Fixed laser bolt collider detection so that enemies can shoot each
other, but not themselves (to prevent scenario of enemies killing
themselves when targeting a player that goes behind it)

Lowered boss2 and boss3 rate of fire.

Added screen fading to the start and end of each scene.

Added PlayerPrefs to carry certain values from one level to the next,
such as Score, level, and playerHealth.

Boss3 has script BossPatrol.cs which gives it random patrol movements in
the play space.

Boss3 also has script smartMissile.cs that allows it to shoot homing
missiles that follow the player.

Made SpawnWaves coroutine more modular by removing the while loop that
checks if it's gameOver and putting that check into Update().
SpawnWaves now only has the sole function of instantiating enemies &
bosses.